tak update, delete and complete

+++ src/routes.js
@@ -1,7 +1,8 @@
 import { Database } from "./database.js";
 import { randomUUID } from 'node:crypto'
 import { buildRoutePath } from "./utils/build-route-path.js";
-import { taskRequest } from "./requests/task-request.js";
+import { taskStoreRequest } from "./requests/task-store-request.js";
+import { taskUpdateRequest } from "./requests/task-update-request.js";
 
 const database = new Database;
 
@@ -26,7 +27,7 @@ export const routes = [
     method: 'POST',
     path: buildRoutePath('/tasks'),
     handler: async (req, res) => {
-      taskRequest(req, res)
+      taskStoreRequest(req, res)
 
       const { title, description, completed_at } = req.body
 
@@ -50,6 +51,12 @@ export const routes = [
     handler: async (req, res) => {
       const { id } = req.params
 
+      const task = database.show('tasks', id)
+
+      if (!task?.id) {
+        return res.writeHead(400).end(JSON.stringify({ message: 'Tarefa não encontrada' }));
+      }
+
       database.delete('tasks', id)
 
       return res.writeHead(204).end()
@@ -60,11 +67,63 @@ export const routes = [
     path: buildRoutePath('/tasks/:id'),
     handler: async (req, res) => {
       const { id } = req.params
-      const { name, email } = req.body
 
-      database.update('tasks', id, { name, email })
+      taskUpdateRequest(req, res, id)
+
+      const task = database.show('tasks', id)
+
+      if (!task?.id) {
+        return res.writeHead(400).end(JSON.stringify({ message: 'Tarefa não encontrada' }));
+      }
+
+      const { title, description } = req.body
+      let data = {}
+
+      if (title) {
+        data.title = title
+      }
+      if (description) {
+        data.description = description
+      }
+
+      database.update('tasks', id, data)
 
       return res.writeHead(204).end()
     }
-  }
+  },
+  {
+    method: 'GET',
+    path: buildRoutePath('/tasks/:id'),
+    handler: async (req, res) => {
+      const { id } = req.params
+
+      const task = database.show('tasks', id)
+
+      return res.end(JSON.stringify(task));
+    }
+  },
+  {
+    method: 'PATCH',
+    path: buildRoutePath('/tasks/:id/complete'),
+    handler: async (req, res) => {
+      const { id } = req.params
+
+
+      const task = database.show('tasks', id)
+
+      if (!task?.id) {
+        return res.writeHead(400).end(JSON.stringify({ message: 'Tarefa não encontrada' }));
+      }
+
+      let completed_at = null;
+
+      if (!task.completed_at) {
+        completed_at = new Date()
+      }
+
+      database.update('tasks', id, { completed_at })
+
+      return res.writeHead(200).end()
+    }
+  },
 ]
